+++ packages/adhere-ui-historyback/lib/historyback.js
@@ -1,2 +1,2 @@
-export default function(e,o){void 0===o&&(o="/"),""!==document.referrer?window.history.go(-1):e&&o&&e.replace(o||"/")}
+export default function(e,r){void 0===r&&(r="/"),!("referrer"in document)||""!==document.referrer?window.history.go(-1):e&&r&&e.replace(r||"/")}
 //# sourceMappingURL=historyback.js.map